streamlit/requirements.txt
streamlit==1.31.0
pandas==2.1.0
plotly==5.18.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
pytz==2024.1
